requirements.txt
cassandra-driver
openai
langflow
